streamline data extraction and analysis process

+++ Sepsis_Bundle/R/util.R
@@ -110,9 +110,7 @@ parse_sql<-function(file_path,...){
       params[params=="&&dbname"]<-"db_name"
       params[params=="&&i2b2"]<-"i2b2_db_schema"
       params[params=="&&PCORNET_CDM"]<-"cdm_db_schema"
-      params[params=="&&start_date"]<-"start_date"      
-      params[params=="&&end_date"]<-"end_date"   
-      params[params=="&&within_d"]<-"within_d"
+      params<-gsub("&&","",params)
     }
     #remove the first line
     line<-gsub("\\t", " ", line)
@@ -234,27 +232,32 @@ drop_tbl<-function(conn,table_name){
   DBMS_type<-attr(conn,"DBMS_type")
   driver_type<-attr(conn,"driver_type")
   
-  if(DBMS_type=="Oracle"){
-    # purge is only required in Oracle for completely destroying temporary tables
-    drop_temp<-paste("drop table",table_name,"purge") 
-    if(driver_type=="OCI"){
-      dbSendQuery(conn,drop_temp)
-    }else if(driver_type=="JDBC"){
-      dbSendUpdate(conn,drop_temp)
+  try_tbl<-try(dbGetQuery(conn,paste("select * from",table_name,"where 1=0")),silent=T)
+  if(is.null(attr(try_tbl,"condition"))){
+    if(DBMS_type=="Oracle"){
+      # purge is only required in Oracle for completely destroying temporary tables
+      drop_temp<-paste("drop table",table_name,"purge") 
+      if(driver_type=="OCI"){
+        dbSendQuery(conn,drop_temp)
+      }else if(driver_type=="JDBC"){
+        dbSendUpdate(conn,drop_temp)
+      }else{
+        stop("Driver type not supported for ",DBMS_type,"!.\n")
+      }
+      
+    }else if(DBMS_type=="tSQL"){
+      drop_temp<-paste("drop table",table_name)
+      if(driver_type=="JDBC"){
+        dbSendUpdate(conn,drop_temp)
+      }else{
+        stop("Driver type not supported for ",DBMS_type,"!.\n")
+      }
+      
     }else{
-      stop("Driver type not supported for ",DBMS_type,"!.\n")
+      warning("DBMS type not supported!")
     }
-    
-  }else if(DBMS_type=="tSQL"){
-    drop_temp<-paste("drop table",table_name)
-    if(driver_type=="JDBC"){
-      dbSendUpdate(conn,drop_temp)
-    }else{
-      stop("Driver type not supported for ",DBMS_type,"!.\n")
-    }
-    
   }else{
-    warning("DBMS type not supported!")
+    cat("table already dropped!")
   }
 }
 
@@ -384,3 +387,144 @@ long_to_sparse_matrix<-function(df,id,variable,val,binary=FALSE){
   return(x_sparse)
 }
 
+get_perf_summ<-function(pred,real,cutoff_fix=NULL,keep_all_cutoffs=F){
+  n<-length(real)
+  
+  # various performace table
+  pred_obj<-ROCR::prediction(pred,real)
+  
+  prc<-performance(pred_obj,"prec","rec")
+  roc<-performance(pred_obj,"sens","spec")
+  nppv<-performance(pred_obj,"ppv","npv")
+  pcfall<-performance(pred_obj,"pcfall")
+  acc<-performance(pred_obj,"acc")
+  fscore<-performance(pred_obj,"f")
+  mcc<-performance(pred_obj,"phi")
+  
+  perf_at<-data.frame(cutoff=prc@alpha.values[[1]],
+                      prec=prc@y.values[[1]],
+                      rec_sens=prc@x.values[[1]],
+                      stringsAsFactors = F) %>% 
+    arrange(cutoff) %>%
+    left_join(data.frame(cutoff=nppv@alpha.values[[1]],
+                         ppv=nppv@y.values[[1]],
+                         npv=nppv@x.values[[1]],
+                         stringsAsFactors = F),
+              by="cutoff") %>%
+    dplyr::mutate(prec_rec_dist=abs(prec-rec_sens)) %>%
+    left_join(data.frame(cutoff=fscore@x.values[[1]],
+                         fscore=fscore@y.values[[1]],
+                         stringsAsFactors = F),
+              by="cutoff") %>%
+    left_join(data.frame(cutoff=roc@alpha.values[[1]],
+                         spec=roc@x.values[[1]],
+                         stringsAsFactors = F),
+              by="cutoff") %>%
+    dplyr::mutate(Euclid_meas=sqrt((1-rec_sens)^2+(0-(1-spec))^2),
+                  Youden_meas=rec_sens+spec-1) %>%
+    left_join(data.frame(cutoff=pcfall@x.values[[1]],
+                         pcfall=pcfall@y.values[[1]],
+                         stringsAsFactors = F),
+              by="cutoff") %>%
+    left_join(data.frame(cutoff=acc@x.values[[1]],
+                         acc=acc@y.values[[1]],
+                         stringsAsFactors = F),
+              by="cutoff") %>%
+    left_join(data.frame(cutoff=mcc@x.values[[1]],
+                         mcc=mcc@y.values[[1]],
+                         stringsAsFactors = F),
+              by="cutoff") %>%
+    filter(prec > 0 & rec_sens > 0 & spec > 0) %>%
+    group_by(cutoff) %>%
+    dplyr::mutate(size=n()) %>%
+    ungroup
+  
+  # performance summary
+  lab1<-pred[real==1]
+  lab0<-pred[real==0]
+  pr<-pr.curve(scores.class0 = lab1,
+               scores.class1 = lab0,curve=F)
+  roc_ci<-pROC::ci.auc(real,pred)
+  
+  perf_summ<-data.frame(overall_meas=c("roauc_low",
+                                       "roauc",
+                                       "roauc_up",
+                                       "opt_thresh",
+                                       "opt_sens",
+                                       "opt_spec",
+                                       "opt_ppv",
+                                       "opt_npv",
+                                       "prauc1",
+                                       "prauc2",
+                                       "opt_prec",
+                                       "opt_rec",
+                                       "opt_fscore"),
+                        meas_val=c(roc_ci[[1]],
+                                   roc_ci[[2]],
+                                   roc_ci[[3]],
+                                   perf_at$cutoff[which.min(perf_at$Euclid_meas)],
+                                   perf_at$rec_sens[which.min(perf_at$Euclid_meas)],
+                                   perf_at$spec[which.min(perf_at$Euclid_meas)],
+                                   perf_at$ppv[which.min(perf_at$Euclid_meas)],
+                                   perf_at$npv[which.min(perf_at$Euclid_meas)],
+                                   pr$auc.integral,
+                                   pr$auc.davis.goadrich,
+                                   perf_at$prec[which.min(perf_at$prec_rec_dist)],
+                                   perf_at$rec_sens[which.min(perf_at$prec_rec_dist)],
+                                   perf_at$fscore[which.min(perf_at$prec_rec_dist)]),
+                        stringsAsFactors = F) %>%
+    bind_rows(perf_at %>% 
+                dplyr::summarize(prec_m=mean(prec,na.rm=T),
+                                 sens_m=mean(rec_sens,na.rm=T),
+                                 spec_m=mean(spec,na.rm=T),
+                                 ppv_m=mean(ppv,na.rm=T),
+                                 npv_m=mean(npv,na.rm=T),
+                                 acc_m=mean(acc,na.rm=T),
+                                 fscore_m=mean(fscore,na.rm=T),
+                                 mcc_m=mean(mcc,na.rm=T)) %>%
+                gather(overall_meas,meas_val))
+  
+  if(!is.null(cutoff_fix)){
+    perf_summ %<>%
+      bind_rows(data.frame(overall_meas=c("sens_fix",
+                                          "spec_fix",
+                                          "ppv_fix",
+                                          "npv_fix",
+                                          "prec_fix",
+                                          "rec_fix",
+                                          "fscore_fix"),
+                           meas_val=c(perf_at$rec_sens[which.min(abs(perf_at$cutoff-cutoff_fix))],
+                                      perf_at$spec[which.min(abs(perf_at$cutoff-cutoff_fix))],
+                                      perf_at$ppv[which.min(abs(perf_at$cutoff-cutoff_fix))],
+                                      perf_at$npv[which.min(abs(perf_at$cutoff-cutoff_fix))],
+                                      perf_at$prec[which.min(abs(perf_at$cutoff-cutoff_fix))],
+                                      perf_at$rec_sens[which.min(abs(perf_at$cutoff-cutoff_fix))],
+                                      perf_at$fscore[which.min(abs(perf_at$cutoff-cutoff_fix))]),
+                           stringsAsFactors = F))
+  }
+  
+  out<-list(perf_summ=perf_summ)
+  if(keep_all_cutoffs){
+    out$perf_at<-perf_at
+  }
+  
+  return(out)
+}
+
+get_calibr<-function(pred,real,n_bin=20){
+  calib<-data.frame(pred=pred,
+                    y=real) %>%
+    arrange(pred) %>%
+    dplyr::mutate(pred_bin = cut(pred,
+                                 breaks=unique(quantile(pred,0:(n_bin)/(n_bin))),
+                                 include.lowest=T,
+                                 labels=F)) %>%
+    ungroup %>% group_by(pred_bin) %>%
+    dplyr::summarize(expos=n(),
+                     bin_lower=min(pred),
+                     bin_upper=max(pred),
+                     bin_mid=median(pred),
+                     Obs = sum(y),
+                     Exp = sum(pred))
+  return(calib)
+}
